ch(refactor): implement central login handler [Starts #46]
