+++ app.js
@@ -17,7 +17,7 @@ const RefreshToken = require('./routes/RefreshToken');
 const FacebookLogin = require('./routes/FacebookLogin');
 
 app.use((req, res, next) => {
-  res.setHeader('Access-Control-Allow-Origin', `${process.env.ACCESS_CONTROL_ALLOW_ORIGIN || '*'}`);
+  res.setHeader('Access-Control-Allow-Origin', process.env.ACCESS_CONTROL_ALLOW_ORIGIN);
   res.setHeader('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content, Accept, Content-Type, Authorization, Set-Cookie');
   res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, PATCH, OPTIONS');
   res.setHeader('Access-Control-Expose-Headers', 'Authorization, X-Refresh-Token');
